Revert "merge jb_mr1_rb1.68 from caf"
This reverts commit 699e5fa3eef8e25866285f24f9add1d044ff59b0.

Conflicts:
	arch/arm/boot/dts/msm8974-clock.dtsi
	arch/arm/mach-msm/Makefile
	arch/arm/mach-msm/Makefile.boot
	arch/arm/mach-msm/include/mach/gpiomux.h
	arch/arm/mach-msm/lpm_levels.c
	drivers/input/touchscreen/Makefile
	drivers/media/platform/msm/camera_v2/Kconfig
	drivers/media/platform/msm/camera_v2/sensor/Makefile
	drivers/media/platform/msm/camera_v2/sensor/ov5648.c
	drivers/power/qpnp-charger.c
	drivers/video/msm/mdss/Kconfig
	drivers/video/msm/mdss/Makefile
	drivers/video/msm/mdss/mdss_dsi.c
	drivers/video/msm/mdss/mdss_dsi_host.c
	drivers/video/msm/mdss/mdss_dsi_panel.c
	drivers/video/msm/mdss/mdss_hdmi_edid.c
	drivers/video/msm/mdss/mdss_mdp_ctl.c
	sound/soc/codecs/wcd9xxx-mbhc.c
	sound/soc/msm/qdsp6v2/msm-dolby-dap-config.c
	sound/soc/msm/qdsp6v2/msm-dolby-dap-config.h

+++ net/nfc/core.c
@@ -185,7 +185,8 @@ int nfc_dep_link_up(struct nfc_dev *dev, int target_index, u8 comm_mode)
 {
 	int rc = 0;
 	u8 *gb;
-	size_t gb_len;
+	u8 gb_len;
+	//size_t gb_len;
 
 	pr_debug("dev_name=%s comm %d\n", dev_name(&dev->dev), comm_mode);
